+++ src-tauri/src/main.rs
@@ -80,8 +80,6 @@ fn start_bot(state: tauri::State<AppState>, app_handle: tauri::AppHandle) {
     let window = app_handle.get_window("client").unwrap();
     let logger = state.logger.clone();
 
-    let _res = window.eval("const overlayElem=document.createElement('div');overlayElem.style.position='absolute',overlayElem.style.left=0,overlayElem.style.top=0,overlayElem.style.height='2px',overlayElem.style.width='2px',overlayElem.style.zIndex=100,overlayElem.style.backgroundColor='red',document.body.appendChild(overlayElem),setInterval(()=>{document.body.style.cursor.indexOf('curattack')>0?overlayElem.style.backgroundColor='green':overlayElem.style.backgroundColor='red'},1)");
-
     let mut image_analyzer: ImageAnalyzer = state.image_analyzer.clone();
 
     image_analyzer.window_id = platform::get_window_id(&window).unwrap_or(0);
@@ -139,6 +137,9 @@ fn start_bot(state: tauri::State<AppState>, app_handle: tauri::AppHandle) {
         let mut shout_behavior = ShoutBehavior::new(&accessor, &logger, &movement);
         let mut last_mode: Option<BotMode> = None;
 
+        // Cursor detection
+        let _res = window.eval("const overlayElem=document.createElement('div');overlayElem.style.position='absolute',overlayElem.style.left=0,overlayElem.style.top=0,overlayElem.style.height='2px',overlayElem.style.width='2px',overlayElem.style.zIndex=100,overlayElem.style.backgroundColor='red',document.body.appendChild(overlayElem),setInterval(()=>{document.body.style.cursor.indexOf('curattack')>0?overlayElem.style.backgroundColor='green':overlayElem.style.backgroundColor='red'},1)");
+
         // Enter main loop
         loop {
             let timer = Timer::start_new("main_loop");
@@ -195,7 +196,7 @@ fn start_bot(state: tauri::State<AppState>, app_handle: tauri::AppHandle) {
             }
 
             // Capture client window
-            image_analyzer.capture_window(&logger);
+            image_analyzer.capture_window(&logger, config.farming_config());
 
             // Try capturing the window contents
             if image_analyzer.image_is_some() {